refactor(frontend): split store/app listeners, helpers, slice types, and PausedBanner
