+++ src/store/skills.ts
@@ -7,7 +7,21 @@ import {
   type Skill,
 } from "@/storage/skills";
 import { detachIdFromAgents } from "./customAgents";
-import type { SkillsSlice, SliceCreator } from "./types";
+import type { SliceCreator } from "./types";
+
+export interface SkillsSlice {
+  /** Shared skills library, mirrored from the `skills` table and ordered
+   *  newest-edited first. Loaded once on init. */
+  skills: Skill[];
+
+  loadSkills: () => Promise<void>;
+  createSkill: (skill: NewSkill) => Promise<Skill>;
+  /** Patch an existing skill; resolves to the merged row, or null if the id is
+   *  unknown. */
+  updateSkill: (id: string, patch: Partial<NewSkill>) => Promise<Skill | null>;
+  /** Delete a skill and detach its id from every custom agent. */
+  deleteSkill: (id: string) => Promise<void>;
+}
 
 // Store slice for the shared skills library. Mirrors the `skills` table
 // (loaded once on init); every mutation writes through to the db and updates
